# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sessionary Backend Server",
    "version": "0.0.0"
  },
  "paths": {
...
    "/lecture": {
      "get": {
        "tags": [
          "lecture"
        ],
        "summary": "Get Lectures",
        "operationId": "get_lectures_lecture_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 10,
              "default": 20,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchRecommendedLecuturesSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "lecture"
        ],
        "summary": "Create Lecture",
        "operationId": "create_lecture_lecture_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLectureBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLectureResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lecture/{lecture_id}": {
      "get": {
        "tags": [
          "lecture"
        ],
        "summary": "Get Lecture",
        "operationId": "get_lecture_lecture__lecture_id__get",
        "parameters": [
          {
            "name": "lecture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lecture Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLectureSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "lecture"
        ],
        "summary": "Update Lecture",
        "operationId": "update_lecture_lecture__lecture_id__patch",
        "parameters": [
          {
            "name": "lecture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lecture Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLectureBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLectureSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ticket/lecture/{lecture_id}": {
      "get": {
        "tags": [
          "ticket"
        ],
        "summary": "Get Lecture Access Status",
        "operationId": "get_lecture_access_status_ticket_lecture__lecture_id__get",
        "parameters": [
          {
            "name": "lecture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lecture Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LectureAccessStatus"
                }
              }
            }
          },
          "401": {
            "description": "Missing token or inactive user."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ticket"
        ],
        "summary": "Use Ticket",
        "operationId": "use_ticket_ticket_lecture__lecture_id__post",
        "parameters": [
          {
            "name": "lecture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lecture Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LectureAccessStatus"
                }
              }
            }
          },
          "401": {
            "description": "Missing token or inactive user."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtistInfoInLecture": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          },
          "is_artist": {
            "type": "boolean",
            "title": "Is Artist"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nickname",
          "is_artist"
        ],
        "title": "ArtistInfoInLecture"
      },
...
      "CreateLectureBody": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": "새로운 강의"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CreateLectureBody"
      },
      "CreateLectureResponseSchema": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LectureDetail"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "CreateLectureResponseSchema"
      },
...
      "FetchRecommendedLecuturesSchema": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/LectureInList"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "FetchRecommendedLecuturesSchema"
      },
...
      "GetLectureSchema": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LectureDetail"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "GetLectureSchema"
      },
      "HTTPValidationError": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LectureAccessStatus": {
        "properties": {
          "accessible": {
            "type": "boolean",
            "title": "Accessible"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "unlimited",
                  "ticket_used",
                  "no_ticket",
                  "ticket_expired"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "ticket_count": {
            "type": "integer",
            "title": "Ticket Count"
          }
        },
        "type": "object",
        "required": [
          "accessible",
          "ticket_count"
        ],
        "title": "LectureAccessStatus"
      },
      "LectureDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtistInfoInLecture"
              },
              {
                "type": "null"
              }
            ]
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonInLecture"
            },
            "type": "array",
            "title": "Lessons"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "thumbnail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail"
          },
          "tags": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "string",
                    "enum": [
                      "원곡카피",
                      "해석버전",
                      "기본기"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "Easy",
                      "Intermediate",
                      "Advanced"
                    ]
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "length_sec": {
            "type": "integer",
            "title": "Length Sec"
          },
          "lecture_count": {
            "type": "integer",
            "title": "Lecture Count"
          },
          "time_created": {
            "type": "string",
            "format": "date-time",
            "title": "Time Created"
          },
          "time_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Time Updated"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LectureProgressData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "artist",
          "lessons",
          "description",
          "thumbnail",
          "tags",
          "length_sec",
          "lecture_count",
          "time_created",
          "time_updated"
        ],
        "title": "LectureDetail"
      },
      "LectureInList": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "thumbnail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtistInfoInLecture"
              },
              {
                "type": "null"
              }
            ]
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonInLecture"
            },
            "type": "array",
            "title": "Lessons"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "tags": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "string",
                    "enum": [
                      "원곡카피",
                      "해석버전",
                      "기본기"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "Easy",
                      "Intermediate",
                      "Advanced"
                    ]
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "length_sec": {
            "type": "integer",
            "title": "Length Sec"
          },
          "lecture_count": {
            "type": "integer",
            "title": "Lecture Count"
          },
          "time_created": {
            "type": "string",
            "format": "date-time",
            "title": "Time Created"
          },
          "time_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Time Updated"
          }
        },
        "type": "object",
        "required": [
          "id",
          "thumbnail",
          "title",
          "artist",
          "lessons",
          "description",
          "tags",
          "length_sec",
          "lecture_count",
          "time_created",
          "time_updated"
        ],
        "title": "LectureInList"
      },
...
      "LectureProgressData": {
        "properties": {
          "completed_count": {
            "type": "integer",
            "title": "Completed Count"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "percent": {
            "type": "integer",
            "title": "Percent"
          },
          "next_lesson_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Lesson Id"
          },
          "completed_lesson_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Completed Lesson Ids"
          }
        },
        "type": "object",
        "required": [
          "completed_count",
          "total_count",
          "percent",
          "next_lesson_id",
          "completed_lesson_ids"
        ],
        "title": "LectureProgressData"
      },
...
      "LessonInLecture": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "length_sec": {
            "type": "integer",
            "title": "Length Sec"
          },
          "lecture_ordering": {
            "type": "integer",
            "title": "Lecture Ordering"
          },
          "time_created": {
            "type": "string",
            "format": "date-time",
            "title": "Time Created"
          },
          "time_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Time Updated"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "length_sec",
          "lecture_ordering",
          "time_created",
          "time_updated"
        ],
        "title": "LessonInLecture"
      },
...
      "PaginationMeta": {
        "properties": {
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "curr_page": {
            "type": "integer",
            "title": "Curr Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "total_items",
          "total_pages",
          "curr_page",
          "per_page"
        ],
        "title": "PaginationMeta"
      },
...
      "UpdateLectureBody": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "string",
                    "enum": [
                      "원곡카피",
                      "해석버전",
                      "기본기"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "Easy",
                      "Intermediate",
                      "Advanced"
                    ]
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "thumbnail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail"
          }
        },
        "type": "object",
        "title": "UpdateLectureBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "satk"
      }
    }
  }
}